Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
sqlalchemy==2.0.35
psycopg==3.2.3
psycopg-binary==3.2.3
alembic==1.13.3

